% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atmle.R
\name{atmle}
\alias{atmle}
\title{Adaptive-TMLE}
\usage{
atmle(
  data,
  S,
  W,
  A,
  Y,
  controls_only,
  family,
  theta_method = "glmnet",
  Pi_method = "glmnet",
  g_method = "glmnet",
  g_delta_method = "glmnet",
  theta_tilde_method = "glmnet",
  bias_working_model = "HAL",
  bias_working_model_formula = NULL,
  pooled_working_model = "HAL",
  pooled_working_model_formula = NULL,
  cross_fit_nuisance = TRUE,
  v_folds = NULL,
  stratify = TRUE,
  g_bounds = NULL,
  Pi_bounds = NULL,
  theta_bounds = NULL,
  target_gwt = TRUE,
  verbose = TRUE,
  max_iter = 50,
  target_method = "relaxed",
  eic_method = "svd_pseudo_inv",
  alpha = 0.05,
  enumerate_basis_args = list(max_degree = 2, smoothness_orders = 1, num_knots = c(20,
    5)),
  browse = FALSE
)
}
\arguments{
\item{data}{A \code{data.frame} containing baseline covariates \eqn{W},
binary treatment indicator \eqn{A} (\eqn{A=1} for active treatment),
outcome \eqn{Y}, and binary study indicator \eqn{S} of whether the
observation is from the randomized controlled trial \eqn{S=1} or from the
external data \eqn{S=0}. If both studies are observational, then \eqn{S=1}
should be the reference study.}

\item{S}{The column name of the \eqn{S} node in \code{data}.}

\item{W}{The column names of the \eqn{W} node in \code{data}.}

\item{A}{The column name of the \eqn{A} node in \code{data}.}

\item{Y}{The column name of the \eqn{Y} node in \code{data}.}

\item{family}{A character string specifying the family of the outcome
\eqn{Y}. Currently only \code{"gaussian"} and \code{"binomial"} are
supported.}

\item{theta_method}{The method to estimate the nuisance function
\eqn{\theta(W,A)=\mathbb{E}(Y\mid W,A)}.
\code{"glm"} for main-term linear model, \code{"glmnet"} for lasso,
\code{"sl3"} for default super learner, or a \code{list} of \code{sl3}
learners. Default is \code{"glmnet"}.}

\item{Pi_method}{The method to estimate the nuisance function
\eqn{\Pi(S\mid W,A)=\mathbb{P}(S=1\mid W,A)}.
\code{"glm"} for main-term linear model, \code{"glmnet"} for lasso,
\code{"sl3"} for default super learner, or a \code{list} of \code{sl3}
learners. Default is \code{"glmnet"}.}

\item{g_method}{The method to estimate the nuisance function
\eqn{g(A\mid W)=\mathbb{P}(A=1\mid W)}.
\code{"glm"} for main-term linear model, \code{"glmnet"} for lasso,
\code{"sl3"} for default super learner, or a \code{list} of \code{sl3}
learners. Default is \code{"glmnet"}.}

\item{g_delta_method}{The method to estimate the nuisance function
\eqn{\tilde{g}_\Delta(1\mid W,A)=\mathbb{P}(\Delta=1\mid W,A)}.
\code{"glm"} for main-term linear model, \code{"glmnet"} for lasso,
\code{"sl3"} for default super learner, or a \code{list} of \code{sl3}
learners. Default is \code{"glmnet"}. \eqn{\Delta} is the indicator of
missing outcome. \code{1} - observed, \code{0} - missing.
Only applicable when there are missing outcomes.}

\item{theta_tilde_method}{The method to estimate the nuisance function
\eqn{\tilde{\theta}(W)=\mathbb{E}(Y\mid W)}.
\code{"glm"} for main-term linear model, \code{"glmnet"} for lasso,
\code{"sl3"} for default super learner, or a \code{list} of \code{sl3}
learners. Default is \code{"glmnet"}.}

\item{bias_working_model}{The working model for the bias estimand.
Either \code{"glmnet"} for main-term lasso-based working model or
\code{"HAL"} for highly adaptive lasso-based working model. Recommended to
use \code{"HAL"}.}

\item{bias_working_model_formula}{An optional formula to specify the
working model for the bias estimand. If provided, this will override the
\code{bias_working_model} argument.}

\item{pooled_working_model}{The working model for the pooled-ATE estimand.
Either \code{"glmnet"} for main-term lasso-based working model or
\code{"HAL"} for highly adaptive lasso-based working model. Recommended to
use \code{"HAL"}.}

\item{pooled_working_model_formula}{An optional formula to specify the
working model for the pooled-ATE estimand. If provided, this will override
the \code{pooled_working_model} argument.}

\item{cross_fit_nuisance}{A logical indicating whether to use cross-fitting
for nuisance function estimation. Default is \code{TRUE}.}

\item{v_folds}{The number of folds for cross-validation (whenever necessary).}

\item{stratify}{A logical indicating whether to stratify the cross-validation
fold regime on binary variables. Default is \code{TRUE}.}

\item{g_bounds}{A numeric vector of lower and upper bounds for the
treatment mechanism. The first element is the lower bound, and the second
element is the upper bound.}

\item{Pi_bounds}{A numeric vector of lower and upper bounds for the
trial enrollment probabilities. The first element is the lower bound,
and the second element is the upper bound.}

\item{theta_bounds}{A numeric vector of lower and upper bounds for the
conditional mean of outcome given baseline covariates and treatment.
The first element is the lower bound, and the second element is the upper
bound.}

\item{target_gwt}{If \code{TRUE}, the treatment mechanism is moved from the
denominator of the clever covariate to the weight when fitting the TMLE
submodel.}

\item{verbose}{A logical indicating whether to print out the progress.
Default is \code{TRUE}.}

\item{max_iter}{Maximum number of iterations for the iterative targeting
procedure. Default is 50.}

\item{target_method}{The targeting method for the working model parameters.
Either \code{"tmle"} for TMLE-type targeting or \code{"relaxed"} for
relaxed targeting. Default is \code{"tmle"}.}

\item{eic_method}{What to do if the information matrix is near singular.
Either \code{"svd_pseudo_inv"} for SVD-based pseudo-inverse, or \code{"diag"}
for adding a small ridge penalty to the diagonal of the information matrix.
Default is \code{"svd_pseudo_inv"}.}

\item{alpha}{Significance level for the confidence interval.
Default is 0.05.}

\item{enumerate_basis_args}{A \code{list} of arguments to be passed to
\code{enumerate_basis} function when constructing the HAL working model
basis.}

\item{browse}{A logical indicating whether to enter the browser for
debugging. Default is \code{FALSE}.}
}
\value{
A \code{list} containing the following elements:
\item{est}{The estimated average treatment effect;}
\item{lower}{The lower bound of the \eqn{(1-\alpha)\%} confidence interval
for the average treatment effect;}
\item{upper}{The upper bound of the \eqn{(1-\alpha)\%} confidence interval
for the average treatment effect;}
\item{psi_pound_est}{The estimated bias parameter;}
\item{psi_pound_lower}{The lower bound of the \eqn{(1-\alpha)\%} confidence
interval for the bias parameter;}
\item{psi_pound_upper}{The upper bound of the \eqn{(1-\alpha)\%} confidence
interval for the bias parameter;}
\item{psi_tilde_est}{The estimated pooled average treatment effect;}
\item{psi_tilde_lower}{The lower bound of the \eqn{(1-\alpha)\%} confidence
interval for the pooled average treatment effect;}
\item{psi_tilde_upper}{The upper bound of the \eqn{(1-\alpha)\%} confidence
interval for the pooled average treatment effect;}
\item{eic}{A numeric vector of the estimated efficient influence curve;}
\item{tau_A}{A \code{list} containing the working model fit for the
pooled-ATE estimand;}
\item{tau_S}{A \code{list} containing the working model fit for the bias
estimand.}
}
\description{
Adaptive-TMLE for estimating the average treatment effect based
on randomized controlled trial augmented with real-world data.
}
