% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonparametric.R
\name{nonparametric}
\alias{nonparametric}
\title{Nonparametric TMLE for Psi (std. over pooled-covariates) and
Psi_2 (std. over RCT-covariates)}
\usage{
nonparametric(
  data,
  S,
  W,
  A,
  Y,
  family,
  Pi_method = "glm",
  g_method = "glm",
  Q_method = "glm",
  Q_pooling = TRUE,
  v_folds = 5,
  verbose = FALSE,
  browse = FALSE
)
}
\description{
Nonparametric TMLE for Psi (std. over pooled-covariates) and
Psi_2 (std. over RCT-covariates)
}
